pyproject.toml
[project]
name = "ai-document-processor"
version = "0.1.0"
description = "An advanced AI-powered document processing web service designed to intelligently extract, analyze, and transform unstructured document content into actionable insights."
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-limiter>=3.9.2",
    "flask-caching>=2.3.0",
    "requests>=2.32.3",
    "pypdf>=3.0.0",
    "openai>=1.58.1",
    "python-dotenv>=1.0.0",
    "pdfminer-six>=20240706",
    "gunicorn>=21.2.0",
    "werkzeug>=3.1.3",
    "langchain>=0.3.13",
    "langchain-community>=0.3.13",
    "faiss-cpu>=1.9.0.post1",
    "sentence-transformers>=3.3.1",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "flask-cors>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "flake8>=7.0.0",
    "pytest>=7.4.4",
]
